+++ service/http/uploads_config.go
@@ -25,7 +25,7 @@ func (cfg *UploadsConfig) TmpDir() string {
 	return os.TempDir()
 }
 
-// Forbid must return true if file extension is not allowed for the upload.
+// Forbids must return true if file extension is not allowed for the upload.
 func (cfg *UploadsConfig) Forbids(filename string) bool {
 	ext := strings.ToLower(path.Ext(filename))