+++ service/static/config.go
@@ -20,7 +20,7 @@ type Config struct {
 	Forbid []string
 }
 
-// Forbid must return true if file extension is not allowed for the upload.
+// Forbids must return true if file extension is not allowed for the upload.
 func (cfg *Config) Forbids(filename string) bool {
 	ext := strings.ToLower(path.Ext(filename))